+++ packages/plugins/Gfi/src/components/Gfi.vue
@@ -1,25 +1,12 @@
 <template>
   <div v-if="renderUi" id="polar-plugin-gfi">
-    <MoveHandle
-      v-if="hasWindowSize && hasSmallWidth"
-      :max-height="maxMobileHeight"
-    >
-      <component
-        :is="contentComponent"
-        :current-properties="currentProperties"
-        :client-width="clientWidth"
-        :export-property="exportProperty"
-        :show-switch-buttons="showSwitchButtons"
-      ></component>
+    <v-card v-if="!windowLayerKeysActive">
+      <v-card-text>{{ $t('common:plugins.gfi.noActiveLayer') }}</v-card-text>
+    </v-card>
+    <MoveHandle v-else-if="renderMoveHandle" :max-height="maxMobileHeight">
+      <component :is="contentComponent" v-bind="contentProps"></component>
     </MoveHandle>
-    <component
-      :is="contentComponent"
-      v-else
-      :current-properties="currentProperties"
-      :client-width="clientWidth"
-      :export-property="exportProperty"
-      :show-switch-buttons="showSwitchButtons"
-    ></component>
+    <component :is="contentComponent" v-else v-bind="contentProps"></component>
   </div>
 </template>
 
@@ -28,10 +15,11 @@ import Vue from 'vue'
 import { GeoJsonProperties } from 'geojson'
 import { mapGetters } from 'vuex'
 import { MoveHandle } from '@polar/components'
-import GfiContent from './GfiContent.vue'
+import Feature from './Feature.vue'
+import List from './List.vue'
 
 export default Vue.extend({
-  name: 'PolarGfi',
+  name: 'GfiPlugin',
   components: {
     MoveHandle,
   },
@@ -41,11 +29,23 @@ export default Vue.extend({
     ...mapGetters('plugin/gfi', [
       'exportPropertyLayerKeys',
       'windowFeatures',
+      'windowLayerKeysActive',
       'visibleWindowFeatureIndex',
       'gfiContentComponent',
+      'showList',
     ]),
     contentComponent(): Vue {
-      return this.gfiContentComponent || GfiContent
+      return this.showList ? List : this.gfiContentComponent || Feature
+    },
+    contentProps(): object {
+      return this.showList
+        ? {}
+        : {
+            currentProperties: this.currentProperties,
+            clientWidth: this.clientWidth,
+            exportProperty: this.exportProperty,
+            showSwitchButtons: this.showSwitchButtons,
+          }
     },
     currentProperties(): GeoJsonProperties {
       const properties = {
@@ -53,7 +53,7 @@ export default Vue.extend({
       }
       const exportProperty =
         this.exportPropertyLayerKeys[properties.polarInternalLayerKey]
-      if (exportProperty.length > 0) {
+      if (exportProperty?.length > 0) {
         delete properties[exportProperty]
       }
       return properties
@@ -71,7 +71,10 @@ export default Vue.extend({
       return ''
     },
     renderUi(): boolean {
-      return this.windowFeatures.length > 0
+      return this.windowFeatures.length > 0 || this.showList
+    },
+    renderMoveHandle(): boolean {
+      return this.hasWindowSize && this.hasSmallWidth
     },
     /** only show switch buttons if multiple property sets are available */
     showSwitchButtons(): boolean {
@@ -112,6 +115,7 @@ export default Vue.extend({
     overflow-x: hidden;
     scrollbar-gutter: stable;
     pointer-events: all;
+    min-width: 300px;
 
     &::v-deep .v-data-table__wrapper {
       /* table cell padding underlaps scrollbar; prevent horizontal scroll */