+++ packages/types/custom/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unpublished
 
+- Feature: Add types for new core state and getters. (`zoomLevel`, `hovered`, `selected`)
+- Feature: Add parameters `renderType`, `featureList`, and `activeLayerPath` to `GfiConfiguration`.
+- Feature: Add type `FeatureList` for `GfiConfiguration`.
 - Feature: Add parameter `renderType` to `GeoLocationConfiguration`.
 - Feature: Add export of `RenderType`.
 - Feature: Add parameter `focusAfterSearch` to `AddressSearchConfiguration`.